Fix "comparison between signed and unsigned" build warning.

+++ expat/xmltok/xmltok.c
@@ -1273,14 +1273,14 @@ XmlInitUnknownEncoding(void *mem,
 		       int (*convert)(void *userData, const char *p),
 		       void *userData)
 {
-  int i;
+  unsigned int i;
   struct unknown_encoding *e = mem;
   for (i = 0; i < sizeof(struct normal_encoding); i++)
     ((char *)mem)[i] = ((char *)&latin1_encoding)[i];
   for (i = 0; i < 128; i++)
     if (latin1_encoding.type[i] != BT_OTHER
         && latin1_encoding.type[i] != BT_NONXML
-	&& table[i] != i)
+	&& table[i] != (int)i)
       return 0;
   for (i = 0; i < 256; i++) {
     int c = table[i];
